+++ src/main/java/qna/service/QnaService.java
@@ -4,8 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qna.CannotDeleteException;
-import qna.NotFoundException;
+import qna.exception.CannotDeleteException;
+import qna.exception.NotFoundException;
 import qna.deletehistory.DeleteHistories;
 import qna.deletehistory.DeleteHistoryRepository;
 import qna.question.Question;
