+++ src/main/java/qna/user/Name.java
@@ -7,6 +7,7 @@ import java.util.Objects;
 @Embeddable
 public class Name {
     private static final int MAX_LENGTH_NAME = 20;
+    private static final String INVALID_NAME_MESSAGE = "사용자 이름은 필수이며 길이는 1이상 20이하여야 합니다.";
 
     @Column(name = "name", nullable = false, length = MAX_LENGTH_NAME)
     private String name;
@@ -21,7 +22,7 @@ public class Name {
 
     private void validateName() {
         if (Objects.isNull(name) || isInvalidNameLength()) {
-            throw new IllegalArgumentException("사용자 이름은 필수이며 길이는 1이상 20이하여야 합니다.");
+            throw new IllegalArgumentException(INVALID_NAME_MESSAGE);
         }
     }
