+++ src/main/java/qna/user/Password.java
@@ -7,6 +7,7 @@ import java.util.Objects;
 @Embeddable
 public class Password {
     private static final int MAX_LENGTH_PASSWORD = 20;
+    private static final String INVALID_PASSWORD_MESSAGE = "비밀번호는 필수이며 길이는 1이상 20이하여야 합니다.";
 
     @Column(name = "password", nullable = false, length = MAX_LENGTH_PASSWORD)
     private String password;
@@ -21,7 +22,7 @@ public class Password {
 
     private void validatePassword() {
         if (Objects.isNull(password) || isInvalidPasswordLength()) {
-            throw new IllegalArgumentException("비밀번호는 필수이며 길이는 1이상 20이하여야 합니다.");
+            throw new IllegalArgumentException(INVALID_PASSWORD_MESSAGE);
         }
     }
