+++ src/main/java/qna/user/UserId.java
@@ -7,6 +7,7 @@ import java.util.Objects;
 @Embeddable
 public class UserId {
     private static final int MAX_LENGTH_USERID = 20;
+    private static final String INVALID_USER_ID_MESSAGE = "사용자 아이디의 필수이며 길이는 1이상 20이하여야 합니다.";
 
     @Column(name = "user_id", nullable = false, length = MAX_LENGTH_USERID, unique = true)
     private String userId;
@@ -21,7 +22,7 @@ public class UserId {
 
     private void validateUserId() {
         if (Objects.isNull(userId) || isInvalidUserIdLength()) {
-            throw new IllegalArgumentException("사용자 아이디의 필수이며 길이는 1이상 20이하여야 합니다.");
+            throw new IllegalArgumentException(INVALID_USER_ID_MESSAGE);
         }
     }
