+++ src/test/java/qna/question/QuestionRepositoryTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import qna.CannotDeleteException;
+import qna.exception.CannotDeleteException;
 import qna.PreExecutionTest;
 import qna.user.User;
 import qna.user.UserId;
