+++ src/test/java/qna/question/QuestionTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import qna.CannotDeleteException;
+import qna.exception.CannotDeleteException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
