+++ src/test/java/qna/user/UserRepositoryTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import qna.UnAuthorizedException;
+import qna.exception.UnAuthorizedException;
 
 import java.util.Optional;
